Added scripts and service units for systemd
Added an update script and a bot executing script compatible with running under systemd.  All these are expected to be under the same directory as the bot's `linuxAIO.sh`.  Made some naming changes and changed the update process as well.  `nadeko-service-create.sh` will create the service units from the directory the script is located in assuming `nadeko-update.sh` and `nadeko.sh` are in that same directory.  Log files will be stored in the same directory as well.